pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-ai-assistant"
version = "2.1.0"
description = "基于 litellm 支持 100+ AI 提供商，通过快捷键在任意界面调用 AI 大模型，CLI + YAML + Web 管理"
authors = [
    { name="mofanx"}
]
license = "MIT"
requires-python = ">=3.8"
readme = "README.md"
dependencies = [
    "keyboard",
    "requests",
    "litellm",
    "pyclip",
    "pyyaml",
    "fastapi",
    "uvicorn",
]

[tool.setuptools.package-data]
ai_assistant = ["default_config.yaml", "web/static/**"]

[project.scripts]
multi_ai_assistant = "ai_assistant.ai_assistant:AI_Assistant"
maa = "ai_assistant.ai_assistant:AI_Assistant"
